+++ src/briefy/leica/config.py
@@ -6,14 +6,28 @@ from prettyconf import config
 NEW_RELIC_LICENSE_KEY = config('NEW_RELIC_LICENSE_KEY', default='')
 
 # Queue
-LEICA_QUEUE = config('LEICA_QUEUE', default='leica-{}'.format(_queue_suffix))
+LEICA_QUEUE = config(
+    'LEICA_QUEUE',
+    default='leica-{env}'.format(env=_queue_suffix)
+)
 
 FILE_PICKER_API_KEY = config('FILE_PICKER_API_KEY', default='')
 
+# Buckets
+UPLOAD_BUCKET = config(
+    'UPLOAD_BUCKET',
+    default='upload-{env}-briefy'.format(env=_queue_suffix)
+)
 
-S3_CREDENTIALS = config('S3_IMAGE_BUCKET', default='')
+IMAGE_BUCKET = config(
+    'IMAGE_BUCKET',
+    default='images-{env}-briefy'.format(env=_queue_suffix)
+)
 
 DATABASE_URL = config('DATABASE_URL',)
 
 # Agoda custom config
 AGODA_DELIVERY_GDRIVE = config('AGODA_DELIVERY_GDRIVE', default='')
+
+# Check if running under DEIS
+DEIS_APP = config('DEIS_APP', default='')